+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.22.0] - 2019-07-16
+
 ### Changed
 
 - Add a new prop that controls if the minicart should show the total quantity of items or just the quantity of different items.